FIX/1. remove references in cipher.py 2. move doctest documentation in the new NeuralNetworkTests class
