+++ claasp/cipher.py
@@ -32,7 +32,6 @@ from claasp.cipher_modules import continuous_tests, code_generator, \
     component_analysis_tests, avalanche_tests, algebraic_tests
 import importlib
 from claasp.cipher_modules.inverse_cipher import *
-from claasp.cipher_modules.neural_network_tests import NeuralNetworkDistinguisher
 
 tii_path = inspect.getfile(claasp)
 tii_dir_path = os.path.dirname(tii_path)
@@ -1233,176 +1232,6 @@ class Cipher:
         return evaluator.evaluate_with_intermediate_outputs_continuous_diffusion_analysis(
             self, cipher_input, sbox_precomputations, sbox_precomputations_mix_columns, verbosity)
 
-    def find_good_input_difference_for_neural_distinguisher(self, difference_positions,
-                                                            initial_population=32, number_of_generations=50,
-                                                            nb_samples=10 ** 4, previous_generation=None,
-                                                            verbose=False):
-        """
-        Return good neural distinguisher input differences for a cipher.
-
-        INPUT:
-
-        - ``difference_positions`` -- **table of booleans**; one for each input to the cipher. True in positions where
-          differences are allowed
-        - ``initial_population`` -- **integer** (default: `32`); parameter of the evolutionary algorithm
-        - ``number_of_generations`` -- **integer** (default: `50`); number of iterations of the evolutionary algorithm
-        - ``nb_samples`` -- **integer** (default: `10`); number of samples for testing each input difference
-        - ``previous_generation`` -- (default: `None`); optional: initial table of differences to try
-        - ``verbose`` -- **boolean** (default: `False`); verbosity
-
-        EXAMPLES::
-
-            sage: from claasp.ciphers.block_ciphers.speck_block_cipher import SpeckBlockCipher
-            sage: cipher = SpeckBlockCipher()
-            sage: diff, scores, highest_round = find_good_input_difference_for_neural_distinguisher(cipher, [True, False], verbose = False, number_of_generations=5)
-        """
-        # from claasp.cipher_modules.neural_network_tests import find_good_input_difference_for_neural_distinguisher
-        return NeuralNetworkDistinguisher(self).find_good_input_difference_for_neural_distinguisher(
-                                                                                        difference_positions,
-                                                                                        initial_population,
-                                                                                        number_of_generations,
-                                                                                        nb_samples,
-                                                                                        previous_generation,
-                                                                                        verbose)
-
-    def train_neural_distinguisher(self, data_generator, starting_round, neural_network, training_samples=10 ** 7,
-                                   testing_samples=10 ** 6, epochs=5, pipeline=True):
-        """
-        Trains a neural distinguisher for the data generated by the data_generator function, using the provided neural network, at round starting_rounds.
-        If pipeline is set to True, retrains the distinguisher for one more round, as long as the validation accuracy remains significant.
-
-        INPUT:
-
-        - ``data_generator`` -- **function**; A dataset generation function, taking as input a cipher (usually self), a number of rounds,
-        and a number of samples, an returns a dataset X, Y, where X is a numpy matrix with one row per sample, and Y is a label veector.
-        To reproduce classical neural distinguisher results, on would use the example below.
-        - ``starting_round`` -- **integer**; number of rounds to analyze
-        - ``neural_network`` -- **(compiled) keras model** (default: `None`); the neural network to use for distinguishing, either a custom one or one
-        returned by the get_neural_network function of neural_network_tests.
-        - ``training_samples`` -- **integer**; (default: `10**7`) number samples used for training
-        - ``testing_samples`` -- **integer**; (default: `10**6`) number samples used for testing
-        - ``pipeline`` -- **boolean**; (default: `True`) If False, only trains for starting_round. If True, increments starting_round and retrain
-        the model as long as the accuracy is statistically significant.
-        - ``verbose`` -- **boolean** (default: `False`); verbosity
-
-        EXAMPLES::
-        sage: from claasp.ciphers.block_ciphers.speck_block_cipher import SpeckBlockCipher
-	    sage: from claasp.cipher_modules.neural_network_tests import get_differential_dataset, get_neural_network
-	    sage: cipher = SpeckBlockCipher()
-	    sage: input_differences = [0x400000, 0]
-	    sage: data_generator = lambda nr, samples: get_differential_dataset(cipher, input_differences, number_of_rounds = nr, samples = samples)
-	    sage: neural_network = get_neural_network('gohr_resnet', input_size = 64)
-	    sage: cipher.train_neural_distinguisher(data_generator, starting_round = 5, neural_network = neural_network)
-        """
-        if pipeline:
-            # from claasp.cipher_modules.neural_network_tests import neural_staged_training
-            acc = NeuralNetworkDistinguisher(self).neural_staged_training(data_generator, starting_round, neural_network, training_samples,
-                                         testing_samples, epochs)
-        else:
-            # from claasp.cipher_modules.neural_network_tests import train_neural_distinguisher
-            acc = NeuralNetworkDistinguisher(self).train_neural_distinguisher(data_generator, starting_round, neural_network, training_samples,
-                                             testing_samples, epochs)
-        return acc
-
-    def train_gohr_neural_distinguisher(self, input_difference, number_of_rounds, depth=1, word_size=0,
-                                        training_samples=10 ** 7, testing_samples=10 ** 6, number_of_epochs=200):
-        """
-          Trains a differential neural distinguisher on nr rounds, for the input difference input_difference, using a slightly
-          modified (AMSGrad instead of cyclic learning rate schedule) depth depth Gohr's RESNet ([Go2019]).
-
-          INPUT:
-
-          - ``input_difference`` -- **list of integers**; The input difference, expressed as a list with one value per
-            input to the cipher.
-          - ``number_of_rounds`` -- **integer**; number of rounds to analyze
-          - ``depth`` -- **integer**; (default: `1`) the depth of the neural network, as defined in Gohr's paper
-          - ``word_size`` -- **integer**; the word size of the cipher, determines the shape of the neural network.
-          Defaults to output_bit_size when unspecified (may reduce the accuracy of the obtained distinguisher).
-          - ``training_samples`` -- **integer**; (default: `10**7`) number samples used for training
-          - ``testing_samples`` -- **integer**; (default: `10**6`) number samples used for testing
-          - ``number_of_epochs`` -- **integer**; (default: `40`) number of training epochs
-
-          EXAMPLES::
-          sage: from claasp.ciphers.block_ciphers.speck_block_cipher import SpeckBlockCipher
-          sage: cipher = SpeckBlockCipher()
-          sage: input_differences = [0x400000, 0]
-          sage: number_of_rounds = 5
-          sage: cipher.train_gohr_neural_distinguisher(input_differences, number_of_rounds, word_size = 16, number_of_epochs = 1)
-          2000/2000 [==============================] - 294s 146ms/step - loss: 0.0890 - acc: 0.8876 - val_loss: 0.0734 - val_acc: 0.9101
-          Validation accuracy at 5 rounds :0.9101160168647766
-          0.9101160168647766
-          """
-
-        def data_generator(nr, samples):
-            return NeuralNetworkDistinguisher(self).get_differential_dataset(input_difference, number_of_rounds=nr,
-                                            samples=samples)
-
-        # from claasp.cipher_modules.neural_network_tests import get_differential_dataset, \
-            # train_neural_distinguisher, get_neural_network
-        input_size = self.output_bit_size * 2
-        neural_network = NeuralNetworkDistinguisher(self).get_neural_network('gohr_resnet', input_size = input_size, depth=depth, word_size=word_size)
-        return NeuralNetworkDistinguisher(self).train_neural_distinguisher(data_generator, number_of_rounds, neural_network, training_samples,
-                                          testing_samples, num_epochs=number_of_epochs)
-
-    def run_autond_pipeline(self, difference_positions=None, optimizer_samples=10 ** 4, optimizer_generations=50,
-                            training_samples=10 ** 7, testing_samples=10 ** 6, number_of_epochs=40, verbose=False, neural_net = 'dbitnet', save_prefix=None):
-        """
-          Runs the AutoND pipeline ([BGHR2023]):
-            - Find an input difference for the inputs set to True in difference_positions using an optimizer
-            - Train a neural distinguisher based on DBitNET for that input difference, increasing the number of rounds
-            until the accuracy is no better than random guessing.
-
-          INPUT:
-
-          - ``difference_positions`` -- **list of booleans**; default: `True in the plaintext position, False in the
-          other positions`. If specified, must have the same length as self.inputs_bit_size, and contain one boolean per
-          input position. The optimizer will look for input differences in the positions set to True; by default,
-          the single-key case will be run.
-          - ``optimizer_samples`` -- **integer**; number of samples used by the optimizer; higher values increase the
-          quality of the optimizer, at the cost of a longer runtime.
-          - ``optimizer_generations`` -- **integer**; (default: `50`) number of generations used by the optimizer;
-          higher values increase the runtime.
-          - ``training_samples`` -- **integer**; (default: `10**7`) number samples used for training
-          - ``testing_samples`` -- **integer**; (default: `10**6`) number samples used for testing
-          - ``number_of_epochs`` -- **integer**; (default: `40`) number of training epochs
-          - ``verbose`` -- **boolean**; (default: `False`) verbosity of the optimizer
-          - ``neural_net`` -- **string**; (default: `dbitnet`) the neural network architecture to use; supports 'dbitnet' and 'gohr_resnet'
-
-
-          EXAMPLES::
-          sage: from claasp.ciphers.block_ciphers.speck_block_cipher import SpeckBlockCipher
-          sage: cipher = SpeckBlockCipher()
-          sage: cipher.run_autond_pipeline()
-          """
-        # from claasp.cipher_modules.neural_network_tests import get_differential_dataset, get_neural_network, \
-        #     int_difference_to_input_differences, neural_staged_training
-
-        def data_generator(nr, samples):
-            return NeuralNetworkDistinguisher(self).get_differential_dataset(input_difference, number_of_rounds=nr,
-                                            samples=samples)
-
-        if difference_positions is None:
-            difference_positions = []
-            for inp in self.inputs:
-                if 'plaintext' in inp:
-                    difference_positions.append(True)
-                else:
-                    difference_positions.append(False)
-        assert True in difference_positions, "At least one position in difference_positions must be set to True. If " \
-                                             "the default value was used, the primitive has no input named `plaintext`."
-
-        diff, scores, highest_round = self.find_good_input_difference_for_neural_distinguisher(difference_positions,
-                                                                                               number_of_generations=optimizer_generations,
-                                                                                               nb_samples=optimizer_samples,
-                                                                                               verbose=verbose)
-        input_difference = NeuralNetworkDistinguisher(self).int_difference_to_input_differences(diff[-1], difference_positions, self.inputs_bit_size)
-        input_size = self.output_bit_size * 2
-        neural_network = NeuralNetworkDistinguisher(self).get_neural_network(neural_net, input_size = input_size)
-        nr = max(1, highest_round-3)
-        print(f'Training {neural_net} on input difference {[hex(x) for x in input_difference]} ({self.inputs}), from round {nr}...')
-        return NeuralNetworkDistinguisher(self).neural_staged_training(data_generator, nr, neural_network, training_samples,
-                                      testing_samples, number_of_epochs) #save_prefix
-
     def generate_bit_based_c_code(self, intermediate_output=False, verbosity=False):
         """
         Return a string containing the C code that defines the self.evaluate() method.
@@ -1779,49 +1608,6 @@ class Cipher:
 
     def make_file_name(self):
         return editor.make_file_name(self._id)
-
-    def neural_network_blackbox_distinguisher_tests(
-            self, nb_samples=10000, hidden_layers=[32, 32, 32], number_of_epochs=10):
-        """
-        Return a python dictionary that contains the accuracies corresponding to each round.
-
-        INPUT:
-
-        - ``nb_samples`` -- **integer** (default: `10000`); how many sample the neural network is trained with
-        - ``hidden_layers`` -- **list** (default: `[32, 32, 32]`); a list containing the number of neurons in each
-          hidden layer of the neural network
-        - ``number_of_epochs`` -- **integer** (default: `10`); how long is the training of the neural network
-
-        EXAMPLES::
-
-            sage: from claasp.ciphers.block_ciphers.speck_block_cipher import SpeckBlockCipher as speck
-            sage: #speck(number_of_rounds=22).neural_network_blackbox_distinguisher_tests(nb_samples = 10) # random
-        """
-        # from claasp.cipher_modules.neural_network_tests import neural_network_blackbox_distinguisher_tests
-        return NeuralNetworkDistinguisher(self).neural_network_blackbox_distinguisher_tests(
-            nb_samples, hidden_layers, number_of_epochs)
-
-    def neural_network_differential_distinguisher_tests(
-            self, nb_samples=10000, hidden_layers=[32, 32, 32], number_of_epochs=10, diff=[0x01]):
-        """
-        Return a python dictionary that contains the accuracies corresponding to each round.
-
-        INPUT:
-
-        - ``nb_samples`` -- **integer** (default: `10000`); how many sample the neural network is trained with
-        - ``hidden_layers`` -- **list** (default: `[32, 32, 32]`); a list containing the number of neurons in each
-          hidden layer of the neural network
-        - ``number_of_epochs`` -- **integer** (default: `10`); how long is the training of the neural network
-        - ``diff`` -- **list** (default: `[0x01]`); list of input differences
-
-        EXAMPLES::
-
-            sage: from claasp.ciphers.block_ciphers.speck_block_cipher import SpeckBlockCipher as speck
-            sage: #speck(number_of_rounds=22).neural_network_differential_distinguisher_tests(nb_samples = 10) # random
-        """
-        # from claasp.cipher_modules.neural_network_tests import neural_network_differential_distinguisher_tests
-        return NeuralNetworkDistinguisher(self).neural_network_differential_distinguisher_tests(
-            nb_samples, hidden_layers, number_of_epochs, diff)
 
     def print(self):
         """
